feat: criando uma classe de erro

+++ igniteteams/src/storage/group/groupCreate.ts
@@ -1,4 +1,5 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AppError } from "@utils/AppError";
 
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 
@@ -7,8 +8,14 @@ import { groupsGetAll } from "./groupsGetAll";
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
-    const storage = JSON.stringify([...storedGroups, newGroup]);
 
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Ja existe um grupo cadastrado com esse nome.");
+    }
+
+    const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
